+++ tests/services/test_auth.py
@@ -2,6 +2,7 @@ import pytest
 from requests import Request
 from tests.conftest import create_mock_response
 
+from py42.services._auth import ApiClientAuth
 from py42.services._auth import BearerAuth
 from py42.services._auth import CustomJWTAuth
 
@@ -17,6 +18,14 @@ def mock_request(mocker):
 def mock_v3_conn(mock_connection, mocker):
     response = create_mock_response(mocker, '{"v3_user_token": "TEST_TOKEN_VALUE"}')
     mock_connection.get.return_value = response
+    return mock_connection
+
+
+@pytest.fixture
+def mock_api_client_conn(mock_connection, mocker):
+    t = '{"access_token": "API_CLIENT_VAL", "token_type": "bearer", "expires_in": 900}'
+    response = create_mock_response(mocker, t)
+    mock_connection.post.return_value = response
     return mock_connection
 
 
@@ -83,3 +92,41 @@ class TestCustomJWTAuth:
         auth = CustomJWTAuth(mock_custom_auth_function)
         jwt_string = auth.get_credentials()
         assert jwt_string == "Bearer token-string"
+
+
+class TestApiClientAuth:
+    def test_call_returns_request_with_expected_header(
+        self, mock_api_client_conn, mock_request
+    ):
+        auth = ApiClientAuth(mock_api_client_conn)
+        request = auth(mock_request)
+        assert request.headers["Authorization"] == "Bearer API_CLIENT_VAL"
+
+    def test_call_only_calls_auth_api_first_time(
+        self, mock_api_client_conn, mock_request
+    ):
+        auth = ApiClientAuth(mock_api_client_conn)
+        auth(mock_request)
+        assert mock_api_client_conn.post.call_count == 1
+        auth(mock_request)
+        assert mock_api_client_conn.post.call_count == 1
+
+    def test_call_calls_auth_api_with_expected_url(
+        self, mock_api_client_conn, mock_request
+    ):
+        auth = ApiClientAuth(mock_api_client_conn)
+        auth(mock_request)
+        params = {"grant_type": "client_credentials"}
+        mock_api_client_conn.post.assert_called_once_with(
+            "/api/v3/oauth/token", params=params
+        )
+
+    def test_clear_credentials_causes_auth_api_to_be_called_on_subsequent_calls(
+        self, mock_api_client_conn, mock_request
+    ):
+        auth = ApiClientAuth(mock_api_client_conn)
+        auth(mock_request)
+        assert mock_api_client_conn.post.call_count == 1
+        auth.clear_credentials()
+        auth(mock_request)
+        assert mock_api_client_conn.post.call_count == 2